api_service/requirements.txt
psycopg2-binary
django
djangorestframework
requests
pyyaml
uritemplate
coreapi